requirements.txt
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
gunicorn
python-dotenv
python-decouple
django-environ